requirements.txt
Flask
flask
numpy
pandas
joblib
scikit-learn
psycopg2-binary
plotly
